+++ app/src/components/market/common/market_top_details_open/index.tsx
@@ -5,7 +5,7 @@ import styled from 'styled-components'
 import { IMPORT_QUESTION_ID_KEY } from '../../../../common/constants'
 import { useConnectedWeb3Context } from '../../../../contexts'
 import { useCompoundService } from '../../../../hooks'
-import { useGraphMarketsFromQuestion } from '../../../../hooks/Graph/useGraphMarketsFromQuestion'
+import { useGraphMarketsFromQuestion } from '../../../../hooks/graph/useGraphMarketsFromQuestion'
 import { useWindowDimensions } from '../../../../hooks/useWindowDimensions'
 import theme from '../../../../theme'
 import { getContractAddress, getNativeAsset, getWrapToken } from '../../../../util/networks'
