+++ app/src/contexts/connectedWeb3.tsx
@@ -3,7 +3,7 @@ import React, { useEffect, useState } from 'react'
 import { useWeb3Context } from 'web3-react'
 
 import { ConnectedBalance, useBalance } from '../hooks'
-import { useCpk } from '../hooks/Cpk/useCpk'
+import { useCpk } from '../hooks/cpk/useCpk'
 import { useSafeApp } from '../hooks/useSafeApp'
 import { CPKService } from '../services'
 import connectors from '../util/connectors'
