+++ app/src/hooks/index.ts
@@ -1,24 +1,24 @@
-export * from './UserTx/useBalance'
-export { use24hsVolume } from './MarketData/use24hsVolume'
+export * from './user_tx/useBalance'
+export { use24hsVolume } from './market_data/use24hsVolume'
 export { useAsyncDerivedValue } from './useAsyncDerivedValue'
-export { useBlockchainMarketMakerData } from './MarketData/useBlockchainMarketMakerData'
+export { useBlockchainMarketMakerData } from './market_data/useBlockchainMarketMakerData'
 export { useCheckContractExists } from './useCheckContractExists'
 export { useCollateral } from './useCollateral'
-export { useCollateralBalance } from './UserTx/useCollateralBalance'
+export { useCollateralBalance } from './user_tx/useCollateralBalance'
 export { useCompoundService } from './useCompoundService'
 export { useContracts } from './useContracts'
-export { useCpkAllowance } from './Cpk/useCpkAllowance'
-export { useCpkProxy } from './Cpk/useCpkProxy'
+export { useCpkAllowance } from './cpk/useCpkAllowance'
+export { useCpkProxy } from './cpk/useCpkProxy'
 export { useDetectAdblocker } from './useDetectAdblocker'
 export { useDocumentDescription } from './useDocumentDescription'
 export { useDocumentTitle } from './useDocumentTitle'
-export { useFundingBalance } from './MarketData/useFundingBalance'
-export { useGraphMarketMakerData } from './Graph/useGraphMarketMakerData'
-export { useGraphMarketUserTxData } from './Graph/useGraphMarketUserTxData'
-export { useGraphMeta } from './Graph/useGraphMeta'
-export { useMarketMakerData } from './MarketData/useMarketMakerData'
+export { useFundingBalance } from './market_data/useFundingBalance'
+export { useGraphMarketMakerData } from './graph/useGraphMarketMakerData'
+export { useGraphMarketUserTxData } from './graph/useGraphMarketUserTxData'
+export { useGraphMeta } from './graph/useGraphMeta'
+export { useMarketMakerData } from './market_data/useMarketMakerData'
 export { usePolling } from './usePolling'
-export { useTokens } from './MarketData/useTokens'
+export { useTokens } from './market_data/useTokens'
 export { useRealityLink } from './useRealityLink'
 export { useSymbol } from './useSymbol'
 export { useMultipleQueries } from './useMultipleQueries'
